Remove all HashiCorp vendored plugins
